Move PatternMatch to the separate file.
